+++ packages/devtools_app/lib/src/screens/performance/panes/timeline_events/perfetto/_perfetto_controller_web.dart
@@ -3,11 +3,10 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-// ignore: avoid_web_libraries_in_flutter, as designed
-import 'dart:html' as html;
 import 'dart:ui_web' as ui_web;
 
 import 'package:flutter/foundation.dart';
+import 'package:web/helpers.dart';
 
 import '../../../../../shared/globals.dart';
 import '../../../../../shared/primitives/trace_event.dart';
@@ -120,8 +119,8 @@ class PerfettoControllerImpl extends PerfettoController {
       return _debugPerfettoUrl;
     }
     final basePath = assetUrlHelper(
-      origin: html.window.location.origin,
-      path: html.window.location.pathname ?? '',
+      origin: window.location.origin,
+      path: window.location.pathname,
     );
     final indexFilePath = ui_web.assetManager
         .getAssetUrl(devToolsExtensionPoints.perfettoIndexLocation);
@@ -129,9 +128,9 @@ class PerfettoControllerImpl extends PerfettoController {
     return '$baseUrl$_embeddedModeQuery';
   }
 
-  html.IFrameElement get perfettoIFrame => _perfettoIFrame;
+  HTMLIFrameElement get perfettoIFrame => _perfettoIFrame;
 
-  late final html.IFrameElement _perfettoIFrame;
+  late final HTMLIFrameElement _perfettoIFrame;
 
   /// The set of trace events that should be shown in the Perfetto trace viewer.
   ///
@@ -166,7 +165,7 @@ class PerfettoControllerImpl extends PerfettoController {
     );
     _initialized = true;
 
-    _perfettoIFrame = html.IFrameElement()
+    _perfettoIFrame = createElementTag('iframe') as HTMLIFrameElement
       // This url is safe because we built it ourselves and it does not include
       // any user input.
       // ignore: unsafe_html