+++ packages/devtools_app/lib/src/shared/config_specific/post_message/post_message.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be found
 // in the LICENSE file.
 
-export 'post_message_stub.dart'
+export '_post_message_stub.dart'
     if (dart.library.js_interop) 'post_message_web.dart';
 
 class PostMessageEvent {