+++ packages/devtools_app_shared/lib/src/service/connected_app.dart
@@ -14,7 +14,7 @@ import 'service_manager.dart';
 final _log = Logger('connected_app');
 
 const flutterLibraryUri = 'package:flutter/src/widgets/binding.dart';
-const dartHtmlLibraryUri = 'dart:html';
+const dartWebLibraryUri = 'dart:js_interop';
 
 // TODO(https://github.com/flutter/devtools/issues/6239): try to remove this.
 @sealed
@@ -81,7 +81,7 @@ class ConnectedApp {
 
   // TODO(kenz): investigate if we can use `libraryUriAvailableNow` instead.
   Future<bool> get isDartWebApp async => _isDartWebApp ??=
-      await serviceManager!.libraryUriAvailable(dartHtmlLibraryUri);
+      await serviceManager!.libraryUriAvailable(dartWebLibraryUri);
 
   bool? get isDartWebAppNow {
     assert(_isDartWebApp != null);