+++ packages/devtools_app_shared/lib/src/ui/theme/_ide_theme_web.dart
@@ -2,11 +2,10 @@
 // Use of this source code is governed by a BSD-style license that can be found
 // in the LICENSE file.
 
-// ignore: avoid_web_libraries_in_flutter, as designed
-import 'dart:html';
 
 import 'package:flutter/widgets.dart';
 import 'package:logging/logging.dart';
+import 'package:web/helpers.dart';
 
 import '../../utils/url/url.dart';
 import '../../utils/utils.dart';