+++ packages/devtools_app_shared/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.0.7
+* Migrate from `dart:html` to `package:web`.
 * Bump the `devtools_shared` dependency to ^5.0.0
 * Remove public getter `libraryRef`, and public methods `getLibrary` and `retrieveFullValueAsString` from `EvalOnDartLibrary`.
 * Change `toString` output for `UnknownEvalException`, `EvalSentinelException`, and `EvalErrorException`.